+++ CHANGELOG.md
@@ -8,6 +8,13 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [1.0.3] - 11-12-2021
+
+### Fixed
+
+- Fixed the `url`, `status`, and `statusText` properties passed into
+  `IntegrationProviderAuthenticationError` in this client
+
 ## 1.0.2 - 3-23-21
 
 ### Added